Consolidate validation within alert component
