+++ apps/frontend/components/common/BrokenReferencesAlert.vue
@@ -1,5 +1,5 @@
 <template>
-  <div v-show="$brokenReferences.warnings.length">
+  <div v-show="referenceErrors.length">
     <VAlert
       class="mt-5 text-left text-subtitle-1"
       color="yellow-lighten-4"
@@ -10,7 +10,7 @@
       <div class="mt-3 text-subtitle-1 font-weight-bold">Detected Issues:</div>
       <div class="pl-1 pr-5 text-subtitle-1">
         <VCard
-          v-for="(warning, index) in $brokenReferences.warnings"
+          v-for="(warning, index) in referenceErrors"
           :key="index"
           class="my-4 pa-4"
           variant="tonal"
@@ -38,7 +38,7 @@
         class="mt-4 ml-1"
         color="yellow-darken-2"
         variant="tonal"
-        @click="repositoryStore.cleanupReferences()"
+        @click="cleanupReferences"
       >
         Remove all broken references
       </VBtn>
@@ -47,9 +47,60 @@
 </template>
 
 <script lang="ts" setup>
+import { schema as schemaConfig } from 'tailor-config-shared';
+
+import { repository as api } from '@/api';
 import { useCurrentRepository } from '@/stores/current-repository';
 
+interface ReferenceError {
+  id?: string;
+  link?: string;
+  message: string;
+}
+
+const { $ceRegistry } = useNuxtApp() as any;
 const repositoryStore = useCurrentRepository();
 
-const { $brokenReferences } = storeToRefs(repositoryStore);
+const errorReport = ref<any>([]);
+const referenceErrors = ref<ReferenceError[]>([]);
+
+const validateReferences = async () => {
+  const { repositoryId } = repositoryStore;
+  if (!repositoryId) throw new Error('Repository not initialized!');
+  const { activities, elements } = await api.validateReferences(repositoryId);
+  referenceErrors.value = [];
+  activities.forEach((it: any) => {
+    referenceErrors.value.push({
+      id: it.src.id,
+      message: `
+        ${it.src.data.name}
+        "${schemaConfig.getLevel(it.src.type).label}" "${it.referenceName}"
+        relationship does not exist anymore. Relationship needs to be removed.
+        If related item removal is published, make sure to publish this
+        item as well after the removal.`,
+    });
+  });
+  elements.forEach((it: any) => {
+    referenceErrors.value.push({
+      link: `/repository/${repositoryId}/editor/${it.src.outlineActivity.id}?elementId=${it.src.uid}`,
+      message: `
+        "${$ceRegistry.get(it.src.type).name}" element relationship
+        "${it.referenceName}" does not exist anymore. Relationship needs to be
+        removed. If related element removal is published, make sure to
+        publish this element as well.`,
+    });
+  });
+  errorReport.value = { activities, elements };
+};
+
+const cleanupReferences = async () => {
+  const { repositoryId } = repositoryStore;
+  if (!repositoryId) throw new Error('Repository not initialized!');
+  await api.cleanupReferences(repositoryId, errorReport.value);
+  await validateReferences();
+};
+
+onMounted(() => {
+  validateReferences();
+});
 </script>
